+++ src/components/BadgerButton/stories.js
@@ -45,12 +45,14 @@ storiesOf('BadgerButton', module)
 					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
 				)}
 				isRepeatable={boolean('repeatable', true)}
+				repeatTimeout={number('repeat timeout (ms)', 4000)}
 				opReturn={array('OP_RETURN', defaultOpReturn)}
 				successFn={() => console.log('success example function called')}
 				failFn={() => console.log('fail example function called')}
 				text={text('Top Text', 'Badger Pay')}
 				showSatoshis={boolean('Toggle Satoshis', true)}
 				showBorder={boolean('Toggle Border', true)}
+				showQR={boolean('Show QR', false)}
 			/>
 		),
 		{
@@ -90,6 +92,22 @@ storiesOf('BadgerButton', module)
 		),
 		{
 			notes: 'Without a text prop, it only shows the price',
+		}
+	)
+	.add(
+		'optional QR code',
+		() => (
+			<BadgerButton
+				amount={0.0001}
+				showQR={boolean('show QR', true)}
+				to={text(
+					'To Address',
+					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
+				)}
+			/>
+		),
+		{
+			notes: 'Optional use a QR code in addition to Button',
 		}
 	)
 	.add(
@@ -160,5 +178,38 @@ storiesOf('BadgerButton', module)
 		),
 		{
 			notes: 'Change the currency and price',
+		}
+	).add(
+		'repeatable payments',
+		() => (
+			<BadgerButton
+				amount={0.0001}
+				to={text(
+					'To Address',
+					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
+				)}
+				isRepeatable={boolean('Repeatable payment', true)}
+				repeatTimeout={number('Reset Timeout (ms)', 5000)}
+			/>
+		),
+		{
+			notes:
+				'Payments which can happen more than once on a single page visit.  Games for example',
+		}
+	).add(
+		'Watch all sources',
+		() => (
+			<BadgerButton
+				amount={0.0001}
+				to={text(
+					'To Address',
+					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
+				)}
+				watchAddress={boolean('watch Address', true)}
+			/>
+		),
+		{
+			notes:
+				'if watchAddress is true, the payment will turn to confirmed when the address receives a payment from any source.  Including other people.  This is ideal to use if the payment codes are unique for the checkout.  Not great if the payment address is shared by users.',
 		}
 	);