+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog - Badger Components React
 
-## 0.2.0 (February xx, 2019)
+## 0.2.0 (February 18, 2019)
 
 Major changes to internal workings of the library.  Can now control amounts independently from Satoshis, should be the base work required to easily add SLP tokens in the near future.